+++ CHANGELOG.md
@@ -6,12 +6,17 @@
 
 *??? ??, ????*
 
+### IMPROVEMENTS
+
+- Update `pkg`: all functions/methods returns errors as global variables
+
 ### CHANGES
 
 - Update `cmd/hidden_lake/messenger`: delete QR codes
 - Update `cmd/hidden_lake/messenger`: append own address in settings
 - Update `cmd/hidden_lake/messenger`: escapeOutput from JS -> html.EscapeString from Go
 - Update `examples/echo_service/default`: middle_hlt -> middle_hls
+- Update `cmd/hidden_lake/applications/messenger`: errors from Fprint -> errors with ErrorPage
 
 ### BUG FIXES